tech(influxdb): refactor and cleanup
