+++ pkg/tsdb/influxdb/query_part_test.go
@@ -7,9 +7,9 @@ import (
 )
 
 func TestInfluxdbQueryPart(t *testing.T) {
-	Convey("Influxdb query part builder", t, func() {
+	Convey("Influxdb query parts", t, func() {
 
-		Convey("should handle field renderer parts", func() {
+		Convey("render field ", func() {
 			part, err := NewQueryPart("field", []string{"value"})
 			So(err, ShouldBeNil)
 
@@ -17,7 +17,7 @@ func TestInfluxdbQueryPart(t *testing.T) {
 			So(res, ShouldEqual, `"value"`)
 		})
 
-		Convey("should handle nested function parts", func() {
+		Convey("render nested part", func() {
 			part, err := NewQueryPart("derivative", []string{"10s"})
 			So(err, ShouldBeNil)
 
@@ -25,7 +25,7 @@ func TestInfluxdbQueryPart(t *testing.T) {
 			So(res, ShouldEqual, "derivative(mean(value), 10s)")
 		})
 
-		Convey("bottom", func() {
+		Convey("render bottom", func() {
 			part, err := NewQueryPart("bottom", []string{"3"})
 			So(err, ShouldBeNil)
 
@@ -33,7 +33,7 @@ func TestInfluxdbQueryPart(t *testing.T) {
 			So(res, ShouldEqual, "bottom(value, 3)")
 		})
 
-		Convey("time", func() {
+		Convey("render time", func() {
 			part, err := NewQueryPart("time", []string{"$interval"})
 			So(err, ShouldBeNil)
 
@@ -41,7 +41,7 @@ func TestInfluxdbQueryPart(t *testing.T) {
 			So(res, ShouldEqual, "time(10s)")
 		})
 
-		Convey("should nest spread function", func() {
+		Convey("render spread", func() {
 			part, err := NewQueryPart("spread", []string{})
 			So(err, ShouldBeNil)
 
@@ -49,7 +49,7 @@ func TestInfluxdbQueryPart(t *testing.T) {
 			So(res, ShouldEqual, `spread(value)`)
 		})
 
-		Convey("should handle suffix parts", func() {
+		Convey("render suffix", func() {
 			part, err := NewQueryPart("math", []string{"/ 100"})
 			So(err, ShouldBeNil)
 
@@ -57,7 +57,7 @@ func TestInfluxdbQueryPart(t *testing.T) {
 			So(res, ShouldEqual, "mean(value) / 100")
 		})
 
-		Convey("should handle alias parts", func() {
+		Convey("render alias", func() {
 			part, err := NewQueryPart("alias", []string{"test"})
 			So(err, ShouldBeNil)
